requirements.txt
# 华尔街母鸡 - 技术指标分析系统依赖

# 基础数据处理
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.12

# 技术指标库
TA-Lib>=0.4.24
talib>=1.0.0

# 数据库连接
redis>=4.5.0
psycopg2-binary>=2.9.0

# 任务调度
schedule>=1.2.0

# 配置管理
python-dotenv>=1.0.0

# HTTP请求
requests>=2.28.0

# 科学计算 (可选)
scipy>=1.10.0
scikit-learn>=1.2.0

# 数据可视化
matplotlib>=3.6.0
plotly>=5.13.0
seaborn>=0.12.0

# 数据验证
pydantic>=1.10.0

# 终端美化
colorama>=0.4.6

# Web框架
flask>=2.2.0
flask-cors>=3.0.0
streamlit>=1.28.0

# 异步处理
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0

# 时区处理
pytz>=2022.7

# 日志处理
loguru>=0.7.0

# 数据库ORM
sqlalchemy>=2.0.0

# 缓存
cachetools>=5.3.0